port the shuffle damage taken chart to core guide code
